Added blog controller and model
