+++ models/blog.js
@@ -1,0 +1,12 @@
+const mongoose = require('mongoose');
+
+const blogSchema = new mongoose.Schema({
+  title: String,
+  author: String,
+  url: String,
+  likes: Number,
+});
+
+const Blog = mongoose.model('Blog', blogSchema);
+
+module.exports = Blog;
